Update README.md
Demand for ”Big Data” expertise started growing over the last few years and so as the demands for the new tools and the techniques that transform this huge amount of data into useful knowledge. The data vol- ume is growing at a rapid rate in every industry or organization. So as ”Big Data” related jobs are in- creasing so rapidly that every sector whether it is Finance, Engineering, Government, Consulting, Scientific, they all are looking for employees who are experts in Data Analysis and can handle ”Big Data”. The tons of data which are generated in these sectors could be structured or unstructured, and they need pro- fessionals who can work on this large quantity of data and turn that data into something valuable and the most crucial of these are ”Data Scientists”. Data Science is gaining traction in every sector and there are di↵erent profiles in Data Science Industry and everyone is an expert in its own field. The people who man- age these profiles have a background in Computer Science or IT because Computer Engineering skills are needed to manage such a large volume of data. This research has been conducted on education needs, roles and responsibilities, skills, certifications and what type of Companies are hiring them. [12].